Move init (->context) module (breaking)
This creates the context module out of the init
module initially defined in commit 20b26a5 as well
as main ctx definition from lib.rs.

src/lib.rs::Pkcs11 -> src/context/mod.rs
src/lib.rs::get_pkcs11 -> src/context/mod.rs
src/types/locking.rs -> src/context/locking.rs
src/types/mod.rs::Info -> src/context/info.rs
src/types/mod.rs::InitializeFlags -> src/context/flags.rs
