+++ cryptoki/src/functions/general_purpose.rs
@@ -4,9 +4,7 @@
 
 use crate::get_pkcs11;
 use crate::types::function::Rv;
-use crate::types::locking::CInitializeArgs;
-use crate::types::Info;
-use crate::Pkcs11;
+use crate::context::{Pkcs11,Info,CInitializeArgs};
 use crate::Result;
 use cryptoki_sys::{CK_C_INITIALIZE_ARGS, CK_INFO};
 use std::ptr;
