+++ cryptoki/src/functions/session_management.rs
@@ -6,7 +6,7 @@ use crate::get_pkcs11;
 use crate::types::function::Rv;
 use crate::session::{Session, SessionInfo, SessionFlags, UserType};
 use crate::slot::Slot;
-use crate::Pkcs11;
+use crate::context::Pkcs11;
 use crate::Result;
 use cryptoki_sys::CK_SESSION_INFO;
 use std::convert::TryInto;
