+++ cryptoki/src/lib.rs
@@ -43,12 +43,6 @@ pub(crate) mod types;
 
 pub use crate::types::function::*;
 
-pub mod init {
-    //! Initialization for the the Pkcs11 context
-    pub use crate::types::Info;
-    pub use crate::types::locking::CInitializeArgs;
-    pub use crate::types::InitializeFlags;
-}
 
 pub mod misc {
     //! Miscellaneous common data types
@@ -58,62 +52,14 @@ pub mod misc {
     pub use crate::types::Bbool;
 }
 
+pub mod context;
 pub mod mechanism;
 pub mod session;
 pub mod object;
 pub mod slot;
 
 use cryptoki_sys::CK_UTF8CHAR;
-use derivative::Derivative;
-use log::error;
 use std::fmt;
-use std::mem;
-use std::path::Path;
-
-/// Directly get the PKCS #11 operation from the context structure and check for null pointers.
-#[macro_export]
-macro_rules! get_pkcs11 {
-    ($pkcs11:expr, $func_name:ident) => {
-        ($pkcs11
-            .function_list
-            .$func_name
-            .ok_or(crate::Error::NullFunctionPointer)?)
-    };
-}
-
-/// Main PKCS11 context. Should usually be unique per application.
-#[derive(Derivative)]
-#[derivative(Debug)]
-pub struct Pkcs11 {
-    // Even if this field is never read, it is needed for the pointers in function_list to remain
-    // valid.
-    #[derivative(Debug = "ignore")]
-    _pkcs11_lib: cryptoki_sys::Pkcs11,
-    function_list: cryptoki_sys::_CK_FUNCTION_LIST,
-}
-
-impl Pkcs11 {
-    /// Instantiate a new context from the path of a PKCS11 dynamic llibrary implementation.
-    pub fn new<P>(filename: P) -> Result<Self>
-    where
-        P: AsRef<Path>,
-    {
-        unsafe {
-            let pkcs11_lib =
-                cryptoki_sys::Pkcs11::new(filename.as_ref()).map_err(Error::LibraryLoading)?;
-            let mut list = mem::MaybeUninit::uninit();
-
-            Rv::from(pkcs11_lib.C_GetFunctionList(list.as_mut_ptr())).into_result()?;
-
-            let list_ptr = *list.as_ptr();
-
-            Ok(Pkcs11 {
-                _pkcs11_lib: pkcs11_lib,
-                function_list: *list_ptr,
-            })
-        }
-    }
-}
 
 #[derive(Debug)]
 /// Main error type
@@ -206,14 +152,6 @@ impl From<std::ffi::NulError> for Error {
 impl From<std::convert::Infallible> for Error {
     fn from(_err: std::convert::Infallible) -> Error {
         unreachable!()
-    }
-}
-
-impl Drop for Pkcs11 {
-    fn drop(&mut self) {
-        if let Err(e) = self.finalize_private() {
-            error!("Failed to finalize: {}", e);
-        }
     }
 }
